Separated Runs info extraction from Cross Valiation Class
