+++ Packages/Classification/ConvNeuralNetwork/lofar_image_creation_test.py
@@ -5,7 +5,7 @@ import os
 
 from Functions.TrainParameters import TrnParamsConvolutional
 from Functions.CrossValidation import NestedCV, SonarRunsCV
-from Functions.NpUtils.DataTransformation import lofar2image
+from Functions.NpUtils.DataTransformation import lofar2image, SonarRunsInfo
 from Functions.SystemIO import exists, mkdir
 
 os.environ['TF_CPP_MIN_LOG_LEVEL'] = '3'
@@ -71,7 +71,7 @@ for cv_name, cv in ncv.cv.items():
                                            ],
                                            loss="categorical_crossentropy")
 
-        cv_info = SonarRunsCV(10, audiodatapath + '/' + database)
+        run_info = SonarRunsInfo(audiodatapath + '/' + database)
 
         def transform_fn(all_data, all_trgt, index_info, info):
             if info == 'train':
@@ -81,10 +81,9 @@ for cv_name, cv in ncv.cv.items():
             return lofar2image(all_data=all_data,
                                all_trgt=all_trgt,
                                index_info=index_info,
-                               class_labels=class_labels,
-                               class_window=image_window,
+                               window_size=image_window,
                                stride=stride,
-                               run_split_info=cv_info)
+                               run_indices_info=run_info)
 
 
         print 'Window size: %i' % image_window
@@ -94,6 +93,6 @@ for cv_name, cv in ncv.cv.items():
             x_test, y_test = transform_fn(all_data=data, all_trgt=trgt,
                                           index_info=test_index, info='val')
             print 'Fold %i:' % fold_count
-            print '\tX: %s \t Trgt: %s' % (x_train.shape, y_train.shape)
-            print '\tX: %s \t Trgt: %s' % (x_test.shape, y_test.shape)
+            print '\tX: %s \t Y: %s' % (x_train.shape, y_train.shape)
+            print '\tX: %s \t Y: %s' % (x_test.shape, y_test.shape)
